remote-robot-controller/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "remote-robot-controller"
version = "1.0.0"
description = "Remote Robot Controller using PyQt5 and MQTT"
authors = [{name = "Robot Team", email = "robot@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
]
dependencies = [
    "PyQt5>=5.15.0",
    "paho-mqtt>=1.6.0",
    "PyYAML>=6.0",
    "Pillow>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-qt>=4.0",
    "black>=22.0",
    "flake8>=4.0",
]

[project.scripts]
robot-controller = "src.main:main"

[project.urls]
Homepage = "https://github.com/robotteam/remote-robot-controller"
Documentation = "https://github.com/robotteam/remote-robot-controller/docs"
Repository = "https://github.com/robotteam/remote-robot-controller.git"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py37']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"